Bring back store to work Add .clang-format

+++ Entities/Tiles/TiledMap.cpp
@@ -57,8 +57,8 @@ namespace Godamn
 	{
 		const auto relPos = ev.getRelativePos(*this);
 
-		const uint8_t tileX = floor(relPos.x / static_cast<float>(m_tileSize.x));
-		const uint8_t tileY = floor(relPos.y / static_cast<float>(m_tileSize.y));
+		const uint8_t tileX = static_cast<uint8_t>(floor(relPos.x / static_cast<float>(m_tileSize.x)));
+		const uint8_t tileY = static_cast<uint8_t>(floor(relPos.y / static_cast<float>(m_tileSize.y)));
 
 		m_tiles[tileX + tileY * m_renderSize.x].onMouseButtonClick(ev);
 		// TODO: redraw only single tile, not whole map (!performance)
